+++ Source/Core/VideoBackends/DX9/TextureCache.cpp
@@ -154,7 +154,7 @@ void TextureCache::TCacheEntry::ReplaceTexture(const u8* src, u32 width, u32 hei
 	s_memPoolTexture[config.pcformat]->GetSurfaceLevel(0, &srcsurf);
 	PDIRECT3DSURFACE9 dstsurface;
 	texture->GetSurfaceLevel(level, &dstsurface);
-	RECT srcr{ 0, 0, width, height };
+	RECT srcr{ 0, 0, LONG(width), LONG(height) };
 	POINT dstp{ 0, 0 };
 	D3D::dev->UpdateSurface(srcsurf, &srcr, dstsurface, &dstp);
 	srcsurf->Release();
